refactor(utils): rename camelCase utility files to kebab-case
Agent-Logs-Url: https://github.com/wolfstar-project/wolfstar.rocks/sessions/f85655d7-47b0-489f-bc04-ea370889f44e
